requirements.txt
Django==1.8.2
django-jinja==1.4.1
gnureadline==6.3.3
ipython==3.1.0
isbnlib==3.5.6
Jinja2==2.7.3
kassis-numbering==0.0.1
lxml==3.4.4
MarkupSafe==0.23
msgpack-python==0.4.6
py-bcrypt==0.4
pylibmc==1.4.3
python3-protobuf==2.5.0
python3-riak-pb==2.0.0.16
pytz==2015.4
regex==2015.5.10
riak==2.2.0
six==1.9.0
